Tolerate 502 errors from GitHub
  HTTP/2 502
  content-type: text/plain; charset=utf-8
  x-served-by: cache-dfw-kdfw8210077
  content-length: 9

+++ unknown-words.sh
@@ -1490,6 +1490,7 @@ call_curl() {
        [ "$response_code" -eq 308 ]; then
       curl_url="$(perl -ne 'next unless /^location:\s*(\S+)/i; print $1' "$response_headers")"
     elif [ "$response_code" -ne 429 ] &&
+       [ "$response_code" -ne 502 ] &&
        [ "$response_code" -ne 503 ]; then
       cat "$response_body"
       rm -f "$response_body"
